requirements.txt
discord.py
openai
psycopg2-binary
pandas
